feat(mcp): port protocol types to dasclaw_mcp + restore McpServerName allowlist (#628 phase 1)
第一阶段：把 MCP 协议的纯类型层迁出 desktop-client/ironclaw，进入工作区共享 crate
`dasclaw_mcp`，同时补回 x-claw fork 在改名过程中丢失的 `McpServerName` 白名单类型。

## 改动范围

新 crate `dasclaw_mcp` v0.0.0-w6（serde / serde_json / thiserror）：
- `protocol`: 753 行纯 serde wire 类型（McpTool / McpRequest / McpResponse / ...）。
  唯一 API 变更：`McpRequest::initialize_with_client_info(id, name, version)`，
  让宿主 crate 决定 wire 上的 clientInfo，避免把 `dasclaw_mcp` 错误地写到协议里。
  保留 `initialize(id)` 作为默认（使用 `env!("CARGO_PKG_NAME"/_VERSION")`），仅用于本 crate 测试。
- `server_name`: 类型化的 `McpServerName(String)`，allowlist `[A-Za-z0-9_.-]`、MAX_LEN=128、
  拒绝 `.` / `..` reserved。13 条单元测试覆盖路径穿越、控制字符、非 ASCII、shell 元字符。

desktop-client/ironclaw 适配：
- `tools/mcp/protocol.rs`: 753 行 → 9 行 re-export shim。
- `tools/mcp/config.rs::validate()`: 用 `McpServerName::new(...)` 替换原本只判空的逻辑；
  新增 `legacy_token_secret_name()`，对照 ironclaw-main 实现 `_` → `-` 的迁移读取。
- `tools/mcp/auth.rs`: `get_access_token` 与 `is_authenticated` 在 NotFound 时回退读旧名。
- `tools/mcp/client.rs:351`: `initialize` 调用改为带入 ironclaw 自身的 CARGO_PKG_NAME/VERSION。

## 非目标（What's NOT in this PR）

- 不迁移 `transport.rs` / `client.rs` / `auth.rs` / `config.rs` 进 `dasclaw_mcp`
  （会拖入 `ToolError` / `JobContext` / `SecretsStore` / `dasclaw_base_dir` 等横切依赖，
  违反 #628 中"禁止补丁式" + "不引入新依赖"两条红线）。这部分留到 F3.2 phase 2+。
- 不接 HookRegistry (ADR-113) / EgressGate (ADR-148) — 待 F3.4/F4 单独 PR。
- 不修复两个**先前就存在**的 clippy 错误
  （`config/job_runtime.rs:278` field_reassign_with_default;
   `tools/builtin/grep_search.rs:372` needless_borrows_for_generic_args）—
  `git diff origin/xClaw -- <files> | wc -l == 0` 已证明非本分支引入。

## 验证

- `cargo check -p dasclaw_mcp --tests`: clean
- `cargo nextest run -p dasclaw_mcp`: 33/33 pass
- `cargo check -p dasclaw --tests`: clean
- `cargo nextest run -p dasclaw`: 4052/4052 pass（22 skipped, 1 leaky）
- `cargo nextest run -p dasclaw --test security_mcp_server_name`: 7/7 pass
- `cargo fmt --all`: OK
- `scripts/check_no_panics.py --base origin/xClaw`: OK
- `scripts/check_no_new_ironclaw_literal.py --base origin/xClaw --head HEAD`: OK
- `cargo clippy --no-deps -p dasclaw_mcp --all-targets -- -D warnings`: clean

## Sources read

- AGENTS.md §"红线" §"测试纪律" §"PR 必填项"
- docs/plans/architecture-refactor/adr-152-*.md §F3.2（MCP port phasing）
- docs/plans/architecture-refactor/adr-112-compatibility-evaluation.md §"非破坏性迁移"
- ironclaw-main/.../tools/mcp/config.rs  §McpServerName 原型
- ironclaw-main/.../tools/mcp/auth.rs    §legacy fallback 读取语义

Refs #628
