+++ crates/dasclaw_mcp/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = "dasclaw_mcp"
-version = "0.0.0-w1"
+version = "0.0.0-w6"
 edition = "2021"
-description = "MCP client with 6 transports (Stdio / SSE / HTTP / WebSocket / SDK / ManagedProxy)."
+description = "Model Context Protocol (MCP) client building blocks shared across dasclaw hosts."
 license = "Apache-2.0 OR MIT"
 publish = false
 
@@ -10,4 +10,6 @@ publish = false
 path = "src/lib.rs"
 
 [dependencies]
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
 thiserror = "1"
